changed model_pao to main model
